calendar problem, continuation of website
-calendar is not working again - unknown reason - tried copying of the
older version but does not work.

-user can register with needed info
the username can be alphanumeric

-the profile page shows the name, surname and email of the user logged
in

-news feed the user can post a comment and the comment will be displayed
with the comment itself, email of the user who commented and the date
and time when the comment was done
the calendar link sends the user to calendar page but again, calendar
not working - seems like ajax link is not working too

-timetable - the same as the old version. did not manage so far

- notes - the subject section does not work, but the user can write a
note with a title and content and this is currently being displayed
ontop of the part where the user writes.

-contact page: send to dropdown lets the user choose between other
accounts and write the message with a null screen when sent
the browse a file, lets the user find the file but cannot send or
discard for now

login/logout working too - login css not fully working