calendar css
previous github commit was about calendar css. it is not detailed as
someone had spilled sprite on my laptop and had to save and backup
really quick just in case it stops working.

calendar css working better. arranged some styling especially when
hovering and the highlighted date and the arrows for next and previous
months.

calendar is almost fully functional as one can enter details and also
rewrite is available. The problem which I am noticing is that when one
changes to a different month, there will be a change in the url (the
year and month are being added together with the function index
therefore when one clicks on the menu to go to another page this is not
allowed, one must go back from the browser itself and then chooses
another page to click on. if the user does not change the month, he/she
can click on another link from the menu without any problems

the notes page is not being found at this stage